[REWORK] Update to the entire system.
The reservation system now uses MongoDb and Express.
It has moved from a SPA to a MPA and supports new features.

The frontend is handled using Sass and Js with AJAX.
The backend is built on Node using Express.
The Database used is MongoDb, which should allow for
great scaling given the ever changing requirements for
the project and data to be handled.

Signed-off-by: Alan Gaurav <alangaurav@outlook.com>